fix: give Grant REP alerts proper spacing
